infra: simplify webpack configuration
I realized that setting server process environment to the one in which
it was built is excessive for the task. The distribution won't be used
directly anyway. Besides it doesn't make sense to start the server on
the port defined in MY .env rather than in the users' one. So while the
approach suggested in [the
article](https://dev.to/sanfra1407/how-to-use-env-file-in-javascript-applications-with-webpack-18df)
is useful in general, it's actually not relevant for this case.